Bug fixes.
Front-end changes:

- Fixed issue in rendering patient and visit details.
- A weird interaction still happens, when moving to another tab, the component gets rerendered when it shouldn't.
- Fixed issues with saving treament branches randomization result.
- Fixed formatting for patient last visit date.

Back-end Changes:
- Added unique restriction on entity `research_participants` so a patient can't be added multiple times to the same research.
- Fixed mapping issues that occured when getting visits for a patient.
- Entity `clinical_visit` now has a foreign key to `research_participants` instead of `participant`.

+++ casciffo-spring-backend/src/main/kotlin/isel/casciffo/casciffospringbackend/aggregates/patients/ResearchPatientsAggregate.kt
@@ -16,7 +16,7 @@ data class ResearchPatientsAggregate(
     var researchPatientId: Int? = null,
 
     //patient info
-    var processId: Int? = null,
+    var processId: Long? = null,
     var fullName: String? = null,
     var gender: String? = null,
     var age: Int? = null,
